requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
requests
schedule
uuid
sqlalchemy
pymysql
python-dotenv
requests
email-validator
apscheduler
imaplib2
pscheduler 
email-validator
psycopg2-binary
streamlit-autorefresh



